+++ packages/next/src/index.ts
@@ -27,6 +27,10 @@ import {
   BuildResultV2Typical as BuildResult,
   BuildResultBuildOutput,
   getInstalledPackageVersion,
+  Span,
+  detectPackageManager,
+  BUILDER_INSTALLER_STEP,
+  BUILDER_COMPILE_STEP,
 } from '@vercel/build-utils';
 import { Route, RouteWithHandle, RouteWithSrc } from '@vercel/routing-utils';
 import {
@@ -199,6 +203,9 @@ function isLegacyNext(nextVersion: string) {
 
 export const build: BuildV2 = async buildOptions => {
   let { workPath, repoRootPath } = buildOptions;
+  const buildSpan =
+    buildOptions.span ?? new Span({ name: BUILDER_COMPILE_STEP });
+
   const {
     files,
     entrypoint,
@@ -336,19 +343,48 @@ export const build: BuildV2 = async buildOptions => {
     await writeNpmRc(entryPath, process.env.NPM_AUTH_TOKEN);
   }
 
-  if (typeof installCommand === 'string') {
-    if (installCommand.trim()) {
-      console.log(`Running "install" command: \`${installCommand}\`...`);
+  const { detectedPackageManager } =
+    detectPackageManager(cliType, lockfileVersion) ?? {};
 
-      await execCommand(installCommand, {
-        ...spawnOpts,
-        cwd: entryPath,
+  const trimmedInstallCommand = installCommand?.trim();
+
+  // If we don't have an install command defined then we fall back
+  // to zero config install, otherwise we run the user generated install command
+  const shouldRunInstallCommand =
+    // Case 1: We have a zero config install
+    typeof trimmedInstallCommand === 'undefined' ||
+    // Case 1: We have a install command which is non zero length
+    Boolean(trimmedInstallCommand);
+
+  buildSpan.setAttributes({
+    install: JSON.stringify(shouldRunInstallCommand),
+  });
+
+  if (shouldRunInstallCommand) {
+    await buildSpan
+      .child(BUILDER_INSTALLER_STEP, {
+        cliType,
+        lockfileVersion: lockfileVersion?.toString(),
+        packageJsonPackageManager,
+        detectedPackageManager,
+        installCommand: trimmedInstallCommand,
+      })
+      .trace(async () => {
+        if (typeof trimmedInstallCommand === 'string') {
+          console.log(
+            `Running "install" command: \`${trimmedInstallCommand}\`...`
+          );
+
+          await execCommand(trimmedInstallCommand, {
+            ...spawnOpts,
+            cwd: entryPath,
+          });
+        } else {
+          await runNpmInstall(entryPath, [], spawnOpts, meta, nodeVersion);
+        }
       });
-    } else {
-      console.log(`Skipping "install" command...`);
-    }
   } else {
-    await runNpmInstall(entryPath, [], spawnOpts, meta, nodeVersion);
+    console.log(`Skipping "install" command...`);
   }
 
   if (spawnOpts.env.VERCEL_ANALYTICS_ID) {
@@ -460,36 +496,43 @@ export const build: BuildV2 = async buildOptions => {
     env.NODE_ENV = 'production';
   }
 
-  if (buildCommand) {
-    // Add `node_modules/.bin` to PATH
-    const nodeBinPaths = getNodeBinPaths({
-      start: entryPath,
-      base: repoRootPath,
-    });
-    const nodeBinPath = nodeBinPaths.join(path.delimiter);
-    env.PATH = `${nodeBinPath}${path.delimiter}${env.PATH}`;
+  await buildSpan
+    .child(BUILDER_COMPILE_STEP, {
+      buildCommand,
+      buildScriptName,
+    })
+    .trace(async () => {
+      if (buildCommand) {
+        // Add `node_modules/.bin` to PATH
+        const nodeBinPaths = getNodeBinPaths({
+          start: entryPath,
+          base: repoRootPath,
+        });
+        const nodeBinPath = nodeBinPaths.join(path.delimiter);
+        env.PATH = `${nodeBinPath}${path.delimiter}${env.PATH}`;
 
-    // Yarn v2 PnP mode may be activated, so force "node-modules" linker style
-    if (!env.YARN_NODE_LINKER) {
-      env.YARN_NODE_LINKER = 'node-modules';
-    }
+        // Yarn v2 PnP mode may be activated, so force "node-modules" linker style
+        if (!env.YARN_NODE_LINKER) {
+          env.YARN_NODE_LINKER = 'node-modules';
+        }
 
-    debug(
-      `Added "${nodeBinPath}" to PATH env because a build command was used.`
-    );
+        debug(
+          `Added "${nodeBinPath}" to PATH env because a build command was used.`
+        );
 
-    console.log(`Running "${buildCommand}"`);
-    await execCommand(buildCommand, {
-      ...spawnOpts,
-      cwd: entryPath,
-      env,
+        console.log(`Running "${buildCommand}"`);
+        await execCommand(buildCommand, {
+          ...spawnOpts,
+          cwd: entryPath,
+          env,
+        });
+      } else if (buildScriptName) {
+        await runPackageJsonScript(entryPath, buildScriptName, {
+          ...spawnOpts,
+          env,
+        });
+      }
     });
-  } else if (buildScriptName) {
-    await runPackageJsonScript(entryPath, buildScriptName, {
-      ...spawnOpts,
-      env,
-    });
-  }
   debug('build command exited');
 
   if (buildCallback) {
@@ -1388,6 +1431,11 @@ export const build: BuildV2 = async buildOptions => {
         requiredServerFilesManifest.config.experimental?.ppr === 'incremental'
       : false;
 
+    const isAppClientSegmentCacheEnabled = requiredServerFilesManifest
+      ? requiredServerFilesManifest.config.experimental?.clientSegmentCache ===
+        true
+      : false;
+
     if (requiredServerFilesManifest) {
       if (!routesManifest) {
         throw new Error(
@@ -1444,6 +1492,7 @@ export const build: BuildV2 = async buildOptions => {
         variantsManifest,
         experimentalPPRRoutes,
         isAppPPREnabled,
+        isAppClientSegmentCacheEnabled,
       });
     }
 
@@ -1964,6 +2013,7 @@ export const build: BuildV2 = async buildOptions => {
       bypassToken: prerenderManifest.bypassToken || '',
       isServerMode,
       isAppPPREnabled: false,
+      isAppClientSegmentCacheEnabled: false,
     }).then(arr =>
       localizeDynamicRoutes(
         arr,
@@ -1994,6 +2044,7 @@ export const build: BuildV2 = async buildOptions => {
         bypassToken: prerenderManifest.bypassToken || '',
         isServerMode,
         isAppPPREnabled: false,
+        isAppClientSegmentCacheEnabled: false,
       }).then(arr =>
         arr.map(route => {
           route.src = route.src.replace('^', `^${dynamicPrefix}`);
@@ -2192,6 +2243,7 @@ export const build: BuildV2 = async buildOptions => {
       isSharedLambdas,
       canUsePreviewMode,
       isAppPPREnabled: false,
+      isAppClientSegmentCacheEnabled: false,
     });
 
     await Promise.all(
@@ -2817,7 +2869,7 @@ export const prepareCache: PrepareCache = async ({
 
   debug('Producing cache file manifest...');
 
-  // for monorepos we want to cache all node_modules
+  // for monorepos we want to cache all node_modules and .yarn/cache
   const isMonorepo = repoRootPath && repoRootPath !== workPath;
   const cacheBasePath = repoRootPath || workPath;
   const cacheEntrypoint = path.relative(cacheBasePath, entryPath);
@@ -2830,6 +2882,12 @@ export const prepareCache: PrepareCache = async ({
     )),
     ...(await glob(
       path.join(cacheEntrypoint, outputDirectory, 'cache/**'),
+      cacheBasePath
+    )),
+    ...(await glob(
+      isMonorepo
+        ? '**/.yarn/cache/**'
+        : path.join(cacheEntrypoint, '.yarn/cache/**'),
       cacheBasePath
     )),
   };