+++ packages/remix/src/build-vite.ts
@@ -1,6 +1,6 @@
 import { readFileSync, promises as fs, statSync, existsSync } from 'fs';
-import { basename, dirname, join, relative, sep } from 'path';
-import { isErrnoException } from '@vercel/error-utils';
+import { basename, dirname, join, relative } from 'path';
+import { isErrnoException, errorToString } from '@vercel/error-utils';
 import { nodeFileTrace, NodeFileTraceOptions } from '@vercel/nft';
 import {
   BuildResultV2Typical,
@@ -13,7 +13,6 @@ import {
   runNpmInstall,
   runPackageJsonScript,
   scanParentDirs,
-  FileBlob,
   FileFsRef,
   EdgeFunction,
   NodejsLambda,
@@ -98,6 +97,17 @@ interface FrameworkSettings {
   buildCommand: string;
   buildResultFilePath: string;
   presetDocumentationLink: string;
+  slug: string;
+  sourceSearchValue: string;
+  edge: {
+    serverSourcePromise: Promise<string>;
+    traceWarningTag: string;
+  };
+  node: {
+    serverSourcePromise: Promise<string>;
+    traceWarningTag: string;
+    options: { useWebApi?: boolean };
+  };
 
   createRenderFunction: (
     options: RenderFunctionOptions
@@ -110,6 +120,17 @@ const REMIX_FRAMEWORK_SETTINGS: FrameworkSettings = {
   buildResultFilePath: '.vercel/remix-build-result.json',
   presetDocumentationLink:
     'https://vercel.com/docs/frameworks/remix#vercel-vite-preset',
+  slug: 'remix',
+  sourceSearchValue: '@remix-run/dev/server-build',
+  edge: {
+    serverSourcePromise: edgeServerSrcPromise,
+    traceWarningTag: '@remix-run/server-runtime',
+  },
+  node: {
+    serverSourcePromise: nodeServerSrcPromise,
+    traceWarningTag: '@remix-run/node',
+    options: {},
+  },
 
   createRenderFunction({
     nodeVersion,
@@ -149,6 +170,18 @@ const REACT_ROUTER_FRAMEWORK_SETTINGS: FrameworkSettings = {
   buildResultFilePath: '.vercel/react-router-build-result.json',
   presetDocumentationLink:
     'https://vercel.com/docs/frameworks/react-router#vercel-react-router-preset',
+  slug: 'react-router',
+  sourceSearchValue: 'virtual:react-router/server-build',
+  // React Router uses the same server source for both node and edge
+  edge: {
+    serverSourcePromise: reactRouterServerSrcPromise,
+    traceWarningTag: 'react-router',
+  },
+  node: {
+    serverSourcePromise: reactRouterServerSrcPromise,
+    traceWarningTag: 'react-router',
+    options: { useWebApi: true },
+  },
 
   createRenderFunction({
     nodeVersion,
@@ -183,6 +216,37 @@ function determineFrameworkSettings(workPath: string) {
     return REACT_ROUTER_FRAMEWORK_SETTINGS;
   }
   return REMIX_FRAMEWORK_SETTINGS;
+}
+
+interface HandlerOptions {
+  rootDir: string;
+  serverBuildPath: string;
+  serverEntryPoint?: string;
+  serverSourcePromise: Promise<string>;
+  sourceSearchValue: string;
+}
+
+async function determineHandler({
+  rootDir,
+  serverBuildPath,
+  serverEntryPoint,
+  serverSourcePromise,
+  sourceSearchValue,
+}: HandlerOptions) {
+  let handler = relative(rootDir, serverBuildPath);
+  let handlerPath = join(rootDir, handler);
+  if (!serverEntryPoint) {
+    const baseServerBuildPath = basename(serverBuildPath, '.js');
+    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
+    handlerPath = join(rootDir, handler);
+
+    const serverSource = await serverSourcePromise;
+    await fs.writeFile(
+      handlerPath,
+      serverSource.replace(sourceSearchValue, `./${baseServerBuildPath}.js`)
+    );
+  }
+  return { handler, handlerPath };
 }
 
 export const build: BuildV2 = async ({
@@ -275,6 +339,33 @@ export const build: BuildV2 = async ({
         cwd: entrypointFsDirname,
       });
     }
+  }
+
+  // If the Build Command or Framework output files according
+  // to the Build Output v3 API, then stop processing here
+  // since the output is already in its final form.
+  const buildOutputPath = join(entrypointFsDirname, '.vercel/output');
+  let buildOutputVersion: undefined | number;
+  try {
+    const boaConfigPath = join(buildOutputPath, 'config.json');
+    const buildResultContents = await fs.readFile(boaConfigPath, 'utf8');
+    buildOutputVersion = JSON.parse(buildResultContents).version;
+  } catch (err: unknown) {
+    if (isErrnoException(err) && err.code === 'ENOENT') {
+      // ENOENT is fine, no need to log
+    } else {
+      debug(
+        `Error reading ".vercel/output/config.json": ${errorToString(err)}`
+      );
+    }
+  }
+  if (buildOutputVersion) {
+    if (buildOutputVersion !== 3) {
+      throw new Error(
+        `The "version" property in ".vercel/output/config.json" must be 3 (received ${buildOutputVersion})`
+      );
+    }
+    return { buildOutputVersion: 3, buildOutputPath };
   }
 
   const buildResultJsonPath = join(
@@ -424,6 +515,51 @@ export const build: BuildV2 = async ({
   return { routes, output, framework: { version: frameworkVersion } };
 };
 
+async function edgeReadFile(fsPath: string) {
+  let source: Buffer | string;
+  try {
+    source = await fs.readFile(fsPath);
+  } catch (err: any) {
+    if (err.code === 'ENOENT' || err.code === 'EISDIR') {
+      return null;
+    }
+    throw err;
+  }
+  if (basename(fsPath) === 'package.json') {
+    // For Edge Functions, patch "main" field to prefer "browser" or "module"
+    const pkgJson = JSON.parse(source.toString());
+
+    for (const prop of ['browser', 'module']) {
+      const val = pkgJson[prop];
+      if (typeof val === 'string') {
+        pkgJson.main = val;
+
+        // Return the modified `package.json` to nft
+        source = JSON.stringify(pkgJson);
+        break;
+      }
+    }
+  }
+  return source;
+}
+
+const EDGE_TRACE_CONDITIONS = [
+  'edge-light',
+  'browser',
+  'module',
+  'import',
+  'require',
+];
+
+const COMMON_NODE_FUNCTION_OPTIONS = {
+  shouldAddHelpers: false,
+  shouldAddSourcemapSupport: false,
+  operationType: 'SSR',
+  supportsResponseStreaming: true,
+} as const;
+
+const COMMON_EDGE_FUNCTION_OPTIONS = { deploymentTarget: 'v8-worker' } as const;
+
 async function createRenderReactRouterFunction(
   nodeVersion: NodeVersion,
   entrypointDir: string,
@@ -434,58 +570,24 @@ async function createRenderReactRouterFunction(
   config: /*TODO: ResolvedNodeRouteConfig*/ any
 ): Promise<EdgeFunction | NodejsLambda> {
   const isEdgeFunction = config.runtime === 'edge';
-  const files: Files = {};
 
-  let handler = relative(rootDir, serverBuildPath);
-  let handlerPath = join(rootDir, handler);
-  if (!serverEntryPoint) {
-    const baseServerBuildPath = basename(serverBuildPath, '.js');
-    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
-    handlerPath = join(rootDir, handler);
-
-    // Copy the `server-react-router.mjs` file into the "build" directory
-    const reactRouterServerSrc = await reactRouterServerSrcPromise;
-    await fs.writeFile(
-      handlerPath,
-      reactRouterServerSrc.replace(
-        'ENTRYPOINT_PLACEHOLDER',
-        `./${baseServerBuildPath}.js`
-      )
-    );
-  }
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise:
+      // React Router has the same promise for both edge and node
+      // so this chooses edge out of convenience
+      REACT_ROUTER_FRAMEWORK_SETTINGS.edge.serverSourcePromise,
+    sourceSearchValue: REACT_ROUTER_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
 
   // Trace the handler with `@vercel/nft`
   let conditions: NodeFileTraceOptions['conditions'];
   let readFile: NodeFileTraceOptions['readFile'];
   if (isEdgeFunction) {
-    conditions = ['edge-light', 'browser', 'module', 'import', 'require'];
-    readFile = async fsPath => {
-      let source: Buffer | string;
-      try {
-        source = await fs.readFile(fsPath);
-      } catch (err: any) {
-        if (err.code === 'ENOENT' || err.code === 'EISDIR') {
-          return null;
-        }
-        throw err;
-      }
-      if (basename(fsPath) === 'package.json') {
-        // For Edge Functions, patch "main" field to prefer "browser" or "module"
-        const pkgJson = JSON.parse(source.toString());
-
-        for (const prop of ['browser', 'module']) {
-          const val = pkgJson[prop];
-          if (typeof val === 'string') {
-            pkgJson.main = val;
-
-            // Return the modified `package.json` to nft
-            source = JSON.stringify(pkgJson);
-            break;
-          }
-        }
-      }
-      return source;
-    };
+    conditions = EDGE_TRACE_CONDITIONS;
+    readFile = edgeReadFile;
   }
   const trace = await nodeFileTrace([handlerPath], {
     base: rootDir,
@@ -494,39 +596,39 @@ async function createRenderReactRouterFunction(
     readFile,
   });
 
-  logNftWarnings(trace.warnings, 'react-router');
+  // React Router has the same warning tag for both edge and node
+  // so this chooses edge out of convenience
+  logNftWarnings(
+    trace.warnings,
+    REACT_ROUTER_FRAMEWORK_SETTINGS.edge.traceWarningTag
+  );
 
-  for (const file of trace.fileList) {
-    files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
-  }
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
 
   let fn: NodejsLambda | EdgeFunction;
   if (isEdgeFunction) {
     fn = new EdgeFunction({
+      ...COMMON_EDGE_FUNCTION_OPTIONS,
       files,
-      deploymentTarget: 'v8-worker',
       entrypoint: handler,
       regions: config.regions,
       framework: {
-        slug: 'react-router',
+        slug: REACT_ROUTER_FRAMEWORK_SETTINGS.slug,
         version: frameworkVersion,
       },
     });
   } else {
     fn = new NodejsLambda({
+      ...COMMON_NODE_FUNCTION_OPTIONS,
       files,
       handler,
       runtime: nodeVersion.runtime,
-      shouldAddHelpers: false,
-      shouldAddSourcemapSupport: false,
-      operationType: 'SSR',
-      supportsResponseStreaming: true,
-      useWebApi: true,
+      useWebApi: REACT_ROUTER_FRAMEWORK_SETTINGS.node.options.useWebApi,
       regions: config.regions,
       memory: config.memory,
       maxDuration: config.maxDuration,
       framework: {
-        slug: 'react-router',
+        slug: REACT_ROUTER_FRAMEWORK_SETTINGS.slug,
         version: frameworkVersion,
       },
     });
@@ -544,25 +646,13 @@ async function createRenderNodeFunction(
   frameworkVersion: string,
   config: /*TODO: ResolvedNodeRouteConfig*/ any
 ): Promise<NodejsLambda> {
-  const files: Files = {};
-
-  let handler = relative(rootDir, serverBuildPath);
-  let handlerPath = join(rootDir, handler);
-  if (!serverEntryPoint) {
-    const baseServerBuildPath = basename(serverBuildPath, '.js');
-    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
-    handlerPath = join(rootDir, handler);
-
-    // Copy the `server-node.mjs` file into the "build" directory
-    const nodeServerSrc = await nodeServerSrcPromise;
-    await fs.writeFile(
-      handlerPath,
-      nodeServerSrc.replace(
-        '@remix-run/dev/server-build',
-        `./${baseServerBuildPath}.js`
-      )
-    );
-  }
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise: REMIX_FRAMEWORK_SETTINGS.node.serverSourcePromise,
+    sourceSearchValue: REMIX_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
 
   // Trace the handler with `@vercel/nft`
   const trace = await nodeFileTrace([handlerPath], {
@@ -570,25 +660,21 @@ async function createRenderNodeFunction(
     processCwd: entrypointDir,
   });
 
-  logNftWarnings(trace.warnings, '@remix-run/node');
+  logNftWarnings(trace.warnings, REMIX_FRAMEWORK_SETTINGS.node.traceWarningTag);
 
-  for (const file of trace.fileList) {
-    files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
-  }
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
 
   const fn = new NodejsLambda({
+    ...COMMON_NODE_FUNCTION_OPTIONS,
     files,
     handler,
     runtime: nodeVersion.runtime,
-    shouldAddHelpers: false,
-    shouldAddSourcemapSupport: false,
-    operationType: 'SSR',
-    supportsResponseStreaming: true,
     regions: config.regions,
     memory: config.memory,
     maxDuration: config.maxDuration,
+    useWebApi: REMIX_FRAMEWORK_SETTINGS.node.options.useWebApi,
     framework: {
-      slug: 'remix',
+      slug: REMIX_FRAMEWORK_SETTINGS.slug,
       version: frameworkVersion,
     },
   });
@@ -604,86 +690,50 @@ async function createRenderEdgeFunction(
   frameworkVersion: string,
   config: /* TODO: ResolvedEdgeRouteConfig*/ any
 ): Promise<EdgeFunction> {
-  const files: Files = {};
-
-  let handler = relative(rootDir, serverBuildPath);
-  let handlerPath = join(rootDir, handler);
-  if (!serverEntryPoint) {
-    const baseServerBuildPath = basename(serverBuildPath, '.js');
-    handler = join(dirname(handler), `server-${baseServerBuildPath}.mjs`);
-    handlerPath = join(rootDir, handler);
-
-    // Copy the `server-edge.mjs` file into the "build" directory
-    const edgeServerSrc = await edgeServerSrcPromise;
-    await fs.writeFile(
-      handlerPath,
-      edgeServerSrc.replace(
-        '@remix-run/dev/server-build',
-        `./${baseServerBuildPath}.js`
-      )
-    );
-  }
-
-  let remixRunVercelPkgJson: string | undefined;
+  const { handler, handlerPath } = await determineHandler({
+    rootDir,
+    serverBuildPath,
+    serverEntryPoint,
+    serverSourcePromise: REMIX_FRAMEWORK_SETTINGS.edge.serverSourcePromise,
+    sourceSearchValue: REMIX_FRAMEWORK_SETTINGS.sourceSearchValue,
+  });
 
   // Trace the handler with `@vercel/nft`
   const trace = await nodeFileTrace([handlerPath], {
     base: rootDir,
     processCwd: entrypointDir,
-    conditions: ['edge-light', 'browser', 'module', 'import', 'require'],
-    async readFile(fsPath) {
-      let source: Buffer | string;
-      try {
-        source = await fs.readFile(fsPath);
-      } catch (err: any) {
-        if (err.code === 'ENOENT' || err.code === 'EISDIR') {
-          return null;
-        }
-        throw err;
-      }
-      if (basename(fsPath) === 'package.json') {
-        // For Edge Functions, patch "main" field to prefer "browser" or "module"
-        const pkgJson = JSON.parse(source.toString());
-
-        for (const prop of ['browser', 'module']) {
-          const val = pkgJson[prop];
-          if (typeof val === 'string') {
-            pkgJson.main = val;
-
-            // Return the modified `package.json` to nft
-            source = JSON.stringify(pkgJson);
-            break;
-          }
-        }
-      }
-      return source;
-    },
+    conditions: EDGE_TRACE_CONDITIONS,
+    readFile: edgeReadFile,
   });
 
-  logNftWarnings(trace.warnings, '@remix-run/server-runtime');
+  logNftWarnings(trace.warnings, REMIX_FRAMEWORK_SETTINGS.edge.traceWarningTag);
 
-  for (const file of trace.fileList) {
-    if (
-      remixRunVercelPkgJson &&
-      file.endsWith(`@remix-run${sep}vercel${sep}package.json`)
-    ) {
-      // Use the modified `@remix-run/vercel` package.json which contains "browser" field
-      files[file] = new FileBlob({ data: remixRunVercelPkgJson });
-    } else {
-      files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
-    }
-  }
+  const files = await getFilesFromTrace({ fileList: trace.fileList, rootDir });
 
   const fn = new EdgeFunction({
+    ...COMMON_EDGE_FUNCTION_OPTIONS,
     files,
-    deploymentTarget: 'v8-worker',
     entrypoint: handler,
     regions: config.regions,
     framework: {
-      slug: 'remix',
+      slug: REMIX_FRAMEWORK_SETTINGS.slug,
       version: frameworkVersion,
     },
   });
 
   return fn;
+}
+
+async function getFilesFromTrace({
+  fileList,
+  rootDir,
+}: {
+  fileList: Set<string>;
+  rootDir: string;
+}) {
+  const files: Files = {};
+  for (const file of fileList) {
+    files[file] = await FileFsRef.fromFsPath({ fsPath: join(rootDir, file) });
+  }
+  return files;
 }